+++ R/ConnectionRead.R
@@ -16,31 +16,31 @@
 #'   are considered equal.
 #' @param check.names If \code{TRUE}, the default, column names will be
 #'   converted to valid R identifiers.
-#' @param row.names A string or an index specifying the column in the DBMS table 
-#'   to use as \code{row.names} in the output data.frame. Defaults to using the
-#'   \code{row_names} column if present. Set to \code{NULL} to never use
-#'   row names.
 #' @param select.cols  A SQL statement (in the form of a character vector of 
 #'    length 1) giving the columns to select. E.g. "*" selects all columns, 
 #'    "x,y,z" selects three columns named as listed.
+#' @inheritParams query
 #' @export
 #' @examples
 #' con <- dbConnect(SQLite())
 #' dbWriteTable(con, "mtcars", mtcars)
 #' dbReadTable(con, "mtcars")
+#' dbGetQuery(con, "SELECT * FROM mtcars WHERE cyl = 8")
 #' 
 #' # Supress row names
 #' dbReadTable(con, "mtcars", row.names = FALSE)
+#' dbGetQuery(con, "SELECT * FROM mtcars WHERE cyl = 8", row.names = FALSE)
 #' 
 #' dbDisconnect(con)
 setMethod("dbReadTable", c("SQLiteConnection", "character"),
   function(conn, name, row.names = NA, check.names = TRUE, select.cols = "*") {
-    out <- dbGetQuery(conn, paste("SELECT", select.cols, "FROM", name))
+    out <- dbGetQuery(conn, paste("SELECT", select.cols, "FROM", name), 
+      row.names = row.names)
     
     if (check.names) {
       names(out) <- make.names(names(out), unique = TRUE)
     }
-    
-    SQL::columnToRownames(out, row.names)
+  
+    out
   }
 )