+++ NEWS.md
@@ -13,7 +13,7 @@
 
 - `dbFetch()` uses the same row name strategy as `dbReadTable()` (#53).
 
-- Updated to SQLite 3.8.8.2
+- Updated to SQLite 3.9.2
 
 - RSQLite always builds with the included source. This prevent bugs due to 
   API mismatches and considerably simplifies the build process.